fixing bottle neck

+++ train.py
@@ -64,7 +64,7 @@ def train_one_batch(net, data_loader, data_size, optimizer, scheduler, mode, dev
         running_loss += loss.item() * point_sets.size(0)
         predictions = torch.argmax(outputs, 1)
         pred_choice = outputs.data.max(1)[1]
-        correct = pred_choice.eq(target.long().data).cpu().sum()
+        correct = pred_choice.eq(target.data).cpu().sum()
         mean_correct.append(correct.item() / float(points.size()[0]))
 
         accuracy += torch.sum(predictions == target)
